[8.0][UPD] Change field's type of percent from integer to float

+++ cmo_account_report/reports/xlsx_report_withholding_income_tax.py
@@ -36,7 +36,7 @@ class WithholdingIncomeTaxView(models.AbstractModel):
     tax_payer = fields.Char(
         string='Tax Payer',
     )
-    percent = fields.Integer(
+    percent = fields.Float(
         string='Percent',
     )
     base_total = fields.Float(
